Just have a CJS and an ESM version of generating the exportPathMap.  Then we don't have to add that `export {};` statement to all of the MDX plugins (that broke the build)
