+++ src/plugins/headings.tsx
@@ -35,4 +35,3 @@ const headingLinkPlugin = () => (tree) => {
 };
 
 module.exports = [headingLinkPlugin];
-export {};
