+++ src/plugins/import.tsx
@@ -83,4 +83,3 @@ const importPlugin = () => (tree) => {
 };
 
 module.exports = importPlugin;
-export {};
