+++ src/plugins/internal-link.tsx
@@ -51,4 +51,3 @@ const internalLinkPlugin = () => (tree) => {
 };
 
 module.exports = internalLinkPlugin;
-export {};
